first round of entityapi updates
- quote column names and clean up insert/update queries
- replace grn with guid
- streamline table structure

fixes

streamline entity history

move EntitySummary into proto

remove EntitySummary

add guid to json

fix tests

change DB_Uuid to DB_NVarchar

fix folder test

convert interface to any

more cleanup

start entity store under grafana-apiserver dskit target

CRUD working, kind of

rough cut of wiring entity api to kube-apiserver

fake grafana user in context

add key to entity

list working

revert unnecessary changes

move entity storage files to their own package, clean up

use accessor to read/write grafana annotations

implement separate Create and Update functions

+++ pkg/services/store/entity/migrations/migrator.go
@@ -4,23 +4,17 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/grafana/grafana/pkg/infra/db"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
-	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/services/sqlstore/migrator"
 	"github.com/grafana/grafana/pkg/services/sqlstore/session"
+	"github.com/grafana/grafana/pkg/services/store/entity/sqlstash"
+
 	"github.com/grafana/grafana/pkg/setting"
 )
 
-func MigrateEntityStore(xdb db.DB, features featuremgmt.FeatureToggles) error {
+func MigrateEntityStore(db sqlstash.EntityDB, features featuremgmt.FeatureToggles) error {
 	// Skip if feature flag is not enabled
 	if !features.IsEnabledGlobally(featuremgmt.FlagEntityStore) {
-		return nil
-	}
-
-	// Migrations depend on upstream xorm implementations
-	sql, ok := xdb.(*sqlstore.SQLStore)
-	if !ok {
 		return nil
 	}
 
@@ -32,41 +26,51 @@ func MigrateEntityStore(xdb db.DB, features featuremgmt.FeatureToggles) error {
 		return nil
 	}
 
-	marker := "Initialize entity tables (v0)" // changing this key wipe+rewrite everything
-	mg := migrator.NewScopedMigrator(sql.GetEngine(), sql.Cfg, "entity")
+	mg := migrator.NewScopedMigrator(db.GetEngine(), db.GetCfg(), "entity")
 	mg.AddCreateMigration()
-	mg.AddMigration(marker, &migrator.RawSQLMigration{})
-	initEntityTables(mg)
+
+	marker := initEntityTables(mg)
 
 	// While this feature is under development, we can completly wipe and recreate
 	// The initial plan is to keep the source of truth in existing SQL tables, and mirrot it
 	// to a kubernetes model.  Once the kubernetes model needs to be preserved,
 	// this code should be removed
-	log, err := mg.GetMigrationLog()
+	exists, err := db.GetEngine().IsTableExist("entity_migration_log")
 	if err != nil {
 		return err
 	}
-	_, found := log[marker]
-	if !found && len(log) > 0 {
-		// Remove the migration log (and potential other orphan tables)
-		tables := []string{"entity_migration_log"}
-
-		ctx := context.Background()
-		err = sql.GetSqlxSession().WithTransaction(ctx, func(tx *session.SessionTx) error {
-			for _, t := range tables {
-				_, err := tx.Exec(ctx, fmt.Sprintf("DROP TABLE IF EXISTS %s", t))
-				if err != nil {
-					return err
-				}
-			}
-			return nil
-		})
+	if exists {
+		log, err := mg.GetMigrationLog()
 		if err != nil {
 			return err
+		}
+		_, found := log[marker]
+		if !found && len(log) > 0 {
+			// Remove the migration log (and potential other orphan tables)
+			tables := []string{"entity_migration_log"}
+
+			ctx := context.Background()
+			err = db.GetSession().WithTransaction(ctx, func(tx *session.SessionTx) error {
+				for _, t := range tables {
+					_, err := tx.Exec(ctx, fmt.Sprintf("DROP TABLE IF EXISTS %s", t))
+					if err != nil {
+						return err
+					}
+				}
+				return nil
+			})
+			if err != nil {
+				return err
+			}
+
+			// remove old entries from in-memory log
+			for id := range log {
+				mg.RemoveMigrationLogs(id)
+			}
 		}
 	}
 
 	return mg.Start(
 		features.IsEnabledGlobally(featuremgmt.FlagMigrationLocking),
-		sql.GetMigrationLockAttemptTimeout())
+		0)
 }
